all tests passing; cleaning up branch and ready to merge
